20200527-Python-01
#Problem1
Write a recursive function which reverse a string word. You cannot use reverse function of Python. 
You need to design a recursive algorithm and implement it. What is time complexity of your solution?

#Problem2
Write e recursive function to check whether a string word whose length is even, is palindromeor not. 
A palindrome word is a word which is equal to its reversed.