20200530-Python-problem8
- Imagine, each student is represented by the following properties:
o firstName
o lastName
o address
o Year of birth
o gpa
Write a function which receives a list of students as its input parameter and return the
student who has the highest average.
- Hint: Use a dictionary (the properties are the keys and the values are the values of the properties, for
instance, “FirstName”:”Ali”). And the define a list of dictionaries and each dictionary represents a student.