20210603_python-assigmment1-4.py
Problem1
Implement a function which receives a list of numbers as its input parameter and apply 
Selection sort on and sort the list and return it.

Problem2
Implement a function which receives a list of numbers as its input parameter and apply 
Mergesort on and sort the list and return it.

Problem3
Implement a function which receives a list of numbers and a number as its input 
parameters and apply Linear search to check whetherthe given number is inside the 
given list or not. The function will return the index of the number in the list if the number 
exists in the list or will return -1 if the number does not exist in the list.

Problem4
Similar to Problem 3, Implement a function which receives a list of 
numbers and a number as its input parameters and apply Binarysearch 
to check whetherthe given number is inside the given list or not. 

The function will return the index of the number in the list if the number
exists in the list or will return -1 if the number does not exist in the list.