Problem4
Implement a function which receives a list of positive integers as inputparameter.The numbers in the list are only between 1 and 99.The list might have repeated numbers. The methodwill return the number that is repeated the most. If there are multiple numbers that are repeated the most, returns only one of them.