Problem7
Define and implement a function which does Linear Search. This function receives two input parameters, one is a list of integers and the other one is a number to search for. The methodreturns -1(negative one)if the number (the second parameter of the function) does not exist in the list or returnsthe index of the number if the number exists in the list.
If there are more than one occurrence of the number, the function will return the index of the first occurrence