Problem8
Design and implement a methodwhich has one input parameter, called topLimit,which is a number between40 and 100. Also, the method has a second input parameter which is also a positive integer, called number. 
The method should create and returns list which contains all numbers in range 2 (two) to topLimitthat are divisible to number.